travel_planning_agent/pyproject.toml
[project]
name = "travel-planning-agent"
version = "0.1.0"
description = "LangGraph 기반 여행 계획 AI 에이전트"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langgraph-cli[inmem]>=0.1.0",
    "langchain>=0.3.0",
    "langchain-upstage>=0.1.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "faiss-cpu>=1.7.0",
    "requests>=2.31.0",
    "pandas>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "ipython>=8.0.0",
    "langsmith>=0.2.0",
]

[tool.pytest.ini_options]
markers = [
    "langsmith: marks tests that log to LangSmith (deselect with '-m \"not langsmith\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
